requirements.txt
Babel==1.3
CairoSVG==1.0.22
ChatterBot==0.4.9
CherryPy==3.2.4
Cython==0.24.1
Django==1.10
Fabric==1.10.2
Flask==0.10.1
Flask-Admin==1.0.7
Flask-Babel==0.9
Flask-Cache==0.13.1
Flask-DebugToolbar==0.8.0
Flask-Excel==0.0.1
Flask-ExtDirect==0.2
Flask-Login==0.2.7
Flask-Mail==0.9.1
Flask-MongoRest==0.1.1
Flask-PyMongo==0.3.1
Flask-Script==0.6.2
Flask-Views==0.2.1
Flask-WTF==0.9.3
Jinja2==2.8
Markdown==2.6.6
MarkupSafe==0.18
MonthDelta==0.9.1
Pigments==1.6
Pillow==2.8.1
PyYAML==3.11
Pygments==2.1.3
SecretStorage==2.3.0
Unidecode==0.4.19
WTForms==1.0.5
WTForms-JSON==0.2.10
Werkzeug==0.9.4
amqp==1.0.13
anyjson==0.3.3
argparse==1.2.1
args==0.1.0
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
billiard==2.7.3.34
blinker==1.3
boto==2.15.0
boxsdk==1.5.3
cairocffi==0.7.2
celery==3.0.24
certifi==2016.8.31
cffi==1.7.0
click==6.6
clint==0.5.1
contextlib2==0.5.1
coreapi==1.32.3
cryptography==1.4
cycler==0.10.0
decorator==4.0.10
dj-database-url==0.4.1
dj-static==0.0.6
django-braces==1.9.0
django-celery==3.1.17
django-chartit==0.2.5
django-crispy-forms==1.6.0
django-debug-toolbar==1.5
django-filter==0.14.0
django-guardian==1.4.5
django-mongodb-engine==0.6.0
django-toolbelt==0.0.1
djangorestframework==3.4.4
djangotoolbox==1.8.0
djmail==0.3
docopt==0.4.0
docutils==0.12
ecdsa==0.13
enum34==1.1.6
eventlet==0.17.4
flasgger==0.5.13
flask-mongoengine==0.7.0
flask-paginate==0.4.1
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.15.2
fuzzywuzzy==0.11.1
gevent==1.0.2
greenlet==0.4.9
gunicorn==19.3.0
httpie==0.9.6
httplib2==0.9.2
idna==2.1
ipaddress==1.0.16
ipython==5.0.0
ipython-genutils==0.1.0
itsdangerous==0.23
itypes==1.1.0
jdcal==1.2
jsondatabase==0.1.1
jsonschema==2.5.1
keyring==9.3.1
kombu==2.5.16
lxml==3.4.2
mailchimp==2.0.9
matplotlib==1.5.1
meld3==0.6.10
memory-profiler==0.41
mimerender==0.5.3
mistune==0.7.3
mock==1.0.1
mongoengine==0.8.4
nameko==2.3.1
nltk==3.2.1
nose==1.3.7
num2words==0.5.3
numpy==1.11.1
oauthlib==1.1.2
openpyxl==2.2.4
paramiko==1.15.2
path.py==8.2.1
pathlib2==2.1.0
pbr==1.10.0
pdfkit==0.5.0
pdfrw==0.2
pexpect==4.2.0
pickleshare==0.7.3
pip-conflict-checker==0.3
pkginfo==1.3.2
plotly==1.12.5
prompt-toolkit==1.0.5
psutil==2.2.1
psycopg2==2.6.2
ptyprocess==0.5.1
pyOpenSSL==16.1.0
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pyexcel==0.1.4
pyexcel-io==0.0.3
pyexcel-webio==0.0.1
pygal==2.2.3
pymongo==3.3.0
pyparsing==2.1.5
python-dateutil==2.2
python-firebase==1.2
python-mimeparse==0.1.4
python-nvd3==0.14.2
python-slugify==1.1.4
python-twitter==3.1
pytz==2013.8
raven==5.10.2
requests==2.11.1
requests-oauthlib==0.6.2
requests-toolbelt==0.7.0
schedule==0.3.1
selenium==2.48.0
simplegeneric==0.8.1
simplejson==3.3.1
six==1.10.0
slackclient==1.0.1
slacker==0.7.3
speaklater==1.3
sqlparse==0.2.1
static3==0.7.0
suds==0.4
svgwrite==1.1.3
textblob==0.11.1
texttable==0.8.4
tornado==3.1.1
traitlets==4.2.2
twilio==5.5.0
twine==1.8.1
uritemplate==3.0.0
urllib3==1.12
wcwidth==0.1.7
websocket-client==0.37.0
wsgiref==0.1.2
xlwt==1.0.0
yagmail==0.5.140
zbar==0.10
zbarlight==1.0.2
